# services/manager/openapi/v1.yaml
openapi: 3.1.0
info:
  title: SkausWatch Manager API
  description: SkausWatch manager service — users, alerts, threat intelligence, approvals, ENDPOINT agent fleet, SIEM, ASM, CodeScan proxy, S3 malware scanning, and research lookups.
  contact:
    name: Penguin Tech Inc
    email: support@penguintech.io
  license:
    name: AGPL-3.0-only
    identifier: AGPL-3.0-only
  version: '1'
paths:
...
  /api/v1/codescan/status:
    get:
      tags:
      - codescan
      summary: GET /darwin/status — CodeScan service status, proxied verbatim.
      description: |-
        This whole router is a thin authenticated proxy to worker-codescan — only
        the canonical `/codescan/*` paths are documented here (the deprecated
        `/darwin/*` aliases mount the same handlers, see `legacy_router`).
        Response bodies are generic JSON objects (`serde_json::Value`) since the
        wire shape is whatever worker-codescan returns.
      operationId: codescan_status
      responses:
        '200':
          description: Proxied worker-codescan response
          content:
            application/json:
              schema: {}
        '401':
          description: Missing or invalid authorization header
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: CodeScan feature not licensed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Cannot connect to worker-codescan
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '504':
          description: worker-codescan request timed out
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearer_jwt: []
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      description: |-
        Documentation-only mirror of `ApiError`'s bare `{"error": msg}` wire shape
        (`BadRequest`/`Unauthorized`/`Forbidden`/`NotFound`/`Conflict` in this
        service all use it, when their body happens to be bare) — `ApiError`
        itself builds its body with `serde_json::json!` rather than a typed
        struct, so this type exists solely to give `utoipa` something to
        reference in `#[utoipa::path]` `responses(...)` clauses. See
        `docs/v2-port/openapi-pattern.md`.
      required:
      - error
      properties:
        error:
          type: string
          description: Human-readable error message.
...
  securitySchemes:
    bearer_jwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
    endpoint_hmac:
      type: apiKey
      in: header
      name: X-API-Key
...